+++ Project_D_S/PreparationEvent.cpp
@@ -1,7 +1,7 @@
 #include "PreparationEvent.h"
 #include "Company.h"
 
-PreparationEvent::PreparationEvent(char typ, int dist, int lt, int cost, Time ET,int ID,Company * C):Event(ET,ID, C) {
+PreparationEvent::PreparationEvent(char typ, int dist, double lt, int cost, Time ET,int ID,Company * C):Event(ET,ID, C) {
 	setType(typ);
 	setDistance(dist);
 	setLoadTime(lt);
